# Synonymize localized text