--- a/lib/util/termops.js
+++ b/lib/util/termops.js
@@ -215,23 +215,33 @@ function getIndexableText(replacer, doc) {
     var uniqTexts = {};
     var indexableText = [];
     var texts = doc._text.split(',');
-    for (var x = 0; x < texts.length; x++) {
-        // push tokens with replacements
-        var tokens = tokenize(token.replaceToken(replacer, texts[x]));
-        if (!tokens.length) continue;
+    getTokens(texts);
 
-        // push tokens with housenum range token if applicable
-        var range = getHousenumRange(doc);
-        if (range) {
-            var l = range.length;
-            while (l--) {
-                add([range[l]].concat(tokens));
-                add(tokens.concat([range[l]]));
+    var keys = Object.keys(doc);
+    for (var i = 0; i < keys.length; i ++) {
+        if (/_text_\w+/.test(keys[i])) getTokens(doc[keys[i]].split(','));
+    }
+
+    function getTokens(texts) {
+        for (var x = 0; x < texts.length; x++) {
+            // push tokens with replacements
+            var tokens = tokenize(token.replaceToken(replacer, texts[x]));
+            if (!tokens.length) continue;
+
+            // push tokens with housenum range token if applicable
+            var range = getHousenumRange(doc);
+            if (range) {
+                var l = range.length;
+                while (l--) {
+                    add([range[l]].concat(tokens));
+                    add(tokens.concat([range[l]]));
+                }
+            } else {
+                add(tokens);
             }
-        } else {
-            add(tokens);
         }
     }
+
     function add(tokens) {
         var key = tokens.join(' ');
         if (!tokens.length || uniqTexts[key]) return;