--- a/lib/util/termops.js
+++ b/lib/util/termops.js
@@ -219,7 +219,7 @@ function getIndexableText(replacer, doc) {
 
     var keys = Object.keys(doc.properties);
     for (var i = 0; i < keys.length; i ++) {
-        if (/_text_\w+/.test(keys[i])) getTokens(doc.properties[keys[i]].split(','));
+        if (/text_\w+/.test(keys[i])) getTokens(doc.properties[keys[i]].split(','));
     }
 
     function getTokens(texts) {