--- a/lib/util/termops.js
+++ b/lib/util/termops.js
@@ -9,6 +9,7 @@ var idPattern = /^(\S+)\.([0-9]+)$/;
 var token = require('./token');
 var permute = require('./permute');
 var murmur = require('murmur');
+var mun = require('model-un');
 
 module.exports.address = address;
 module.exports.maskAddress = maskAddress;
@@ -219,7 +220,8 @@ function getIndexableText(replacer, doc) {
 
     var keys = Object.keys(doc.properties);
     for (var i = 0; i < keys.length; i ++) {
-        if (/text_\w+/.test(keys[i])) getTokens(doc.properties[keys[i]].split(','));
+        var code = keys[i].match(/text_(.+)/);
+        if (code && mun.hasLanguage(code[1])) getTokens(doc.properties[keys[i]].split(','));
     }
 
     function getTokens(texts) {
@@ -259,7 +261,13 @@ function decollide(replacer, doc, subq) {
     // remove leading/trailing numtokens.
     subq = subq.replace(/^([0-9]*[#]+)/,'').replace(/([0-9]*[#]+)$/,'');
 
-    var texts = token.replaceToken(replacer, doc.properties["carmen:text"]).split(',');
+    var texts = token.replaceToken(replacer, doc.properties['carmen:text']).split(',');
+    var keys = Object.keys(doc.properties);
+    for (var i = 0; i < keys.length; i ++) {
+        var code = keys[i].match(/text_(.+)/);
+        if (code && mun.hasLanguage(code[1])) texts = texts.concat(token.replaceToken(replacer, doc.properties[keys[i]]).split(','));
+    }
+
     var a = texts.length;
     var fails = 0;
     while (a--) {