--- a/lib/util/termops.js
+++ b/lib/util/termops.js
@@ -219,7 +219,8 @@ function getIndexableText(replacer, doc) {
     getTokens(texts);
 
     var keys = Object.keys(doc.properties);
-    for (var i = 0; i < keys.length; i ++) {
+    var length = keys.length;
+    for (var i = 0; i < length; i ++) {
         var code = keys[i].match(/text_(.+)/);
         if (code && mun.hasLanguage(code[1])) getTokens(doc.properties[keys[i]].split(','));
     }
@@ -263,7 +264,8 @@ function decollide(replacer, doc, subq) {
 
     var texts = token.replaceToken(replacer, doc.properties['carmen:text']).split(',');
     var keys = Object.keys(doc.properties);
-    for (var i = 0; i < keys.length; i ++) {
+    var length = keys.length;
+    for (var i = 0; i < length; i ++) {
         var code = keys[i].match(/text_(.+)/);
         if (code && mun.hasLanguage(code[1])) texts = texts.concat(token.replaceToken(replacer, doc.properties[keys[i]]).split(','));
     }