--- a/lib/util/termops.js
+++ b/lib/util/termops.js
@@ -222,7 +222,8 @@ function getIndexableText(replacer, doc) {
     var length = keys.length;
     for (var i = 0; i < length; i ++) {
         var code = keys[i].match(/text_(.+)/);
-        if (code && mun.hasLanguage(code[1])) getTokens(doc.properties[keys[i]].split(','));
+        if (code && mun.hasLanguage(code[1]) &&
+            doc.properties[keys[i]]) getTokens(doc.properties[keys[i]].split(','));
     }
 
     function getTokens(texts) {
@@ -267,7 +268,8 @@ function decollide(replacer, doc, subq) {
     var length = keys.length;
     for (var i = 0; i < length; i ++) {
         var code = keys[i].match(/text_(.+)/);
-        if (code && mun.hasLanguage(code[1])) texts = texts.concat(token.replaceToken(replacer, doc.properties[keys[i]]).split(','));
+        if (code && mun.hasLanguage(code[1]) &&
+            doc.properties[keys[i]]) texts = texts.concat(token.replaceToken(replacer, doc.properties[keys[i]]).split(','));
     }
 
     var a = texts.length;