--- a/package.json
+++ b/package.json
@@ -27,7 +27,8 @@
     "turf-linestring": "1.0.x",
     "turf-point": "2.0.x",
     "turf-point-on-surface": "1.0.x",
-    "unidecode": "0.1.3"
+    "unidecode": "0.1.3",
+    "model-un": "^0.0.3"
   },
   "devDependencies": {
     "retire": "0.4.x",