Additional `carmen:text` parameters (e.g. `:carmen:text_ru` or `carmen:text_es`) for local language support in place names are indexed as synonyms.