@camilleanne 2015-09-18
@yhahn I'd like your thoughts on my modification to decollide. Also, I feel like perhaps the step that's happening in `getIndexableText` and `decollide` may only need to happen once.