@camilleanne 2015-10-02
Benchmark:
This branch (with changes to decollide):
```
# decollide
decollide (collision) x 96,011 ops/sec ±0.72% (91 runs sampled)
decollide (clean) x 101,122 ops/sec ±1.05% (85 runs sampled)
decollide (with localization) x 70,597 ops/sec ±7.10% (55 runs sampled)
decollide (with localization + unidecode) x 35,708 ops/sec ±1.91% (62 runs sampled)
```


Master:
```
# decollide
decollide (collision) x 164,937 ops/sec ±8.41% (61 runs sampled)
decollide (clean) x 138,953 ops/sec ±0.73% (88 runs sampled)
```