@camilleanne 2015-10-05
@apendleton so what do we need to do to get your fork of unidecode running in this?