@apendleton 2015-10-05
@camilleanne should just be swapping out the dep in package.json and changing the require, but I want to do a bit more testing for correctness if we're actually thinking about using it. I've stolen the upstream unidecode tests for now, but they're kind of lacking and I'd like to be more confident that what I have is right before relying on it. I think it's a small-ish lift, but wanted to make sure it was something that would be useful to you before I spent workday time on it. Sounds like that's a yes? If so, I'll test a bit more and submit a PR against this branch with the swap.