@apendleton 2015-10-06
Cool. I spent a bit of time at the end of the day today identifying behavior differences between my code and node-unidecode, and there are some upstream behaviors that I'm 99% sure are wrong. I'm just gonna go for parity for now, but at some point it might be worth figuring out what the actual correct thing is. I think I can get to being confident enough to test with carmen without needing to tackle any of that, though, probably by end of day tomorrow.