@apendleton 2015-10-07
@camilleanne okay, [submitted a PR](https://github.com/mapbox/carmen/pull/345) to switch this branch to unidecode-cxx.

New benchmarks from Travis:

    # decollide
    
    decollide (collision) x 115,361 ops/sec ±0.37% (97 runs sampled)
    decollide (clean) x 123,077 ops/sec ±0.53% (99 runs sampled)
    decollide (with localization) x 84,043 ops/sec ±0.48% (102 runs sampled)
    decollide (with localization + unidecode) x 77,548 ops/sec ±0.09% (103 runs sampled)