@apendleton 2015-10-07
One other thing I should mention: neither old nor new unidecode do anything to detect whether or not running on a string is necessary before running on it. I'm not sure where the unidecode call is going to happen, and specifically whether it's only going to be called on text that we think is non-latin (like, from Chinese data or whatever) or whether it'll be called on everything, but if the latter, we could add a check in c++ land that would return the input if there's no operation necessary. That check would be pretty cheap, but not free (no allocations or function calls, but some bit-twiddling), so probably only worth it if we think there's going to be a decent number of calls where this thing ends up being a no-op.

Or it may be fast enough as-is, and not worth worrying about.