@sbma44 2015-10-07
@apendleton I think that check is worth consideration. our customers are still overwhelmingly anglophone.