@apendleton 2015-10-07
@sbma44 okay, I can cut a ticket. Given that this would probably result in a (slight) speed hit on non-latin text in exchange for a speed win on latin text, I'd probably need a test corpus to benchmark that represents a realistic mix of latin vs. non-latin text to see if the overall speed gain is worth it. Thoughts on what to use?