@sbma44 2015-10-07
@apendleton I think this is a question for #devops. I suspect we can replay some portion of incoming queries. OTOH, our traffic is so bursty that it might take a long time to sample something representative. Curious if @yhahn has ideas here.