@yhahn 2015-10-07
@apendleton @sbma44 I think we can leave this no-op to next time/if it matters. Going from 77k => 84k ops/sec (the fastest it would get us) is not that important at this point : )